refactor: Extract Abbreviations
